pycharm

+++ app/views.py
@@ -2,8 +2,8 @@ from django.shortcuts import render, redirect
 
 from app.models import Printer, GcodeFile, RegisteredPrinter
 from django.http import FileResponse, JsonResponse, HttpResponse, Http404
-
 from app.form import PrinterForm
+from django.forms.models import model_to_dict
 
 
 # Create your views here.t
@@ -11,7 +11,10 @@ from app.form import PrinterForm
 def index(request):
     return render(request, 'index.html')
 
+
 import os
+
+
 def printerlist(request):
     if request.method == 'GET':
         printer = RegisteredPrinter.objects.all()
@@ -25,8 +28,8 @@ def register_printer(request):
         if printer_id:
             i = RegisteredPrinter.objects.filter(printer_id=printer_id).first()
             v = PrinterForm(instance=i, prefix='vv')
+            Printer.objects.get_or_create(printer_id=printer_id)
         else:
-            print("hah")
             v = PrinterForm(prefix='vv')
 
         return render(request, 'register_printer.html', locals())
@@ -34,35 +37,58 @@ def register_printer(request):
     if request.method == 'POST':
         v = PrinterForm(data=request.POST, prefix='vv')
         if v.is_valid():
-            printer_id = v['printer_id']
-            owner = v['owner']
-            address = v['address']
+            printer_id = request.POST['vv-printer_id']
+            print(printer_id)
             result = RegisteredPrinter.objects.filter(printer_id=printer_id)
             if not result:
                 v.save()
+                Printer.objects.get_or_create(printer_id=printer_id)
             return redirect('/list/')
         else:
             error_msg = v.errors.as_json()
             return render(request, 'register_printer.html', locals())
 
+
 def del_printer(request):
     if request.method == 'POST':
-        printer_sel = RegisteredPrinter.objects.filter(printer_id=request.POST['printer_id'])
-        printer_sel.delete()
+        RegisteredPrinter.objects.filter(printer_id=request.POST['printer_id']).delete()
+        Printer.objects.filter(printer_id=request.POST['printer_id']).delete()
         return redirect('/list/')
+
+
 
 def printerlist_api(request):
     printer_list = {}
     if request.method == 'GET':
-        printer_on = Printer.objects.filter(is_ready='True').values('printer_id')
-        printer_off = Printer.objects.filter(is_ready='False').values('printer_id')
-        for pon in printer_on:
-            printer_list[pon['printer_id']] = 'Online'
-        for poff in printer_off:
-            printer_list[poff['printer_id']] = 'Offline'
-        return JsonResponse(printer_list)
+        printer = Printer.objects.all()
+
+        for p in printer:
+            if p.is_ready == 'True':
+                state = '在线，可供打印'
+            elif p.is_printing == 'True':
+                state = '打印中'
+            elif p.is_cancelling == 'True':
+                state = '正在取消'
+            elif p.is_pausing == 'True':
+                state = '暂停中'
+            elif p.is_paused == 'True':
+                state = '已暂停打印'
+            elif p.is_closed_or_error == 'True':
+                state = '打印机关闭或错误'
+            else:
+                state = '离线'
+            printer_list[p.printer_id] = state
+        return JsonResponse(printer_list, json_dumps_params={'ensure_ascii': False})
+
     if request.method == 'POST':
-        pass
+        return JsonResponse(printer_list)
+
+
+def del_alldata(request):
+    Printer.objects.all().delete()
+    RegisteredPrinter.objects.all().delete()
+    GcodeFile.objects.all().delete()
+    return HttpResponse('数据库文件已删除')
 
 
 def delprinterdata(request):
@@ -72,33 +98,29 @@ def delprinterdata(request):
 
 def delgcodedata(request):
     if request.method == 'POST':
-        gcode_file_sel = GcodeFile.objects.filter(gcode_id=request.POST['gcode_id'])
-        gcode_path = GcodeFile.objects.values('gcode_safepath').get(gcode_id=request.POST['gcode_id'])[
-            'gcode_safepath']
-        print(gcode_path)
-        try:
-            os.remove(gcode_path)
-        except Exception:
-            pass
-        gcode_file_sel.delete()
-    return redirect('/list/')
-
-#
-def downloadindex(request):
-    return render(request, 'downloadindex.html')
-
-
-def download1(request):
-    file_path = "/gcodefiles\\Gnomon_Northen.gcode"
-    try:
-        # r=FileResponse(open(file_path, 'rb'))
-        # r['content_type']='application/ocet-stream'
-        # r['content_Disposition'] = 'attachment;filename=Gnomon_Northen.gcode'
-        f = open(file_path, 'rb')
-        r = FileResponse(f, as_attachment=True, filename='Gnomon_Northen.gcode')
-        return r
-    except Exception:
-        raise Http404('download error')
+        if GcodeFile.objects.filter(gcode_id=request.POST['gcode_id']):
+            gcode_file_sel = GcodeFile.objects.get(gcode_id=request.POST['gcode_id'])
+            gcode_path = model_to_dict(gcode_file_sel)['gcode_safepath']
+            try:
+                os.remove(gcode_path)
+            except Exception:
+                return redirect('/list/')
+            gcode_file_sel.delete()
+            return redirect('/list/')
+        else:
+            return redirect('/list/')
+    if request.method == 'GET':
+        gcode_file_sel = GcodeFile.objects.all().first()
+        if gcode_file_sel:
+            gcode_path = model_to_dict(gcode_file_sel)['gcode_safepath']
+            try:
+                os.remove(gcode_path)
+            except Exception:
+                return redirect('/list/')
+            gcode_file_sel.delete()
+            return redirect('/list/')
+        else:
+            return redirect('/list/')
 
 
 def download_gcode_file(request, filename):
@@ -119,6 +141,8 @@ def upload_gcode_file(request):
     if not os.path.exists(gcode_folder):
         os.makedirs("gcodefiles")
 
+    if GcodeFile.objects.all().first():
+        return HttpResponse('已有文件上传，请先打印')
     if request.method == 'POST':
         gcode_file_re = request.FILES.get("gcode_file", None)
         gcode_path = os.path.join(gcode_folder, gcode_file_re.name)
@@ -130,7 +154,7 @@ def upload_gcode_file(request):
 
         GcodeFile.objects.get_or_create(
             gcode_name=gcode_file_re.name,
-            gcode_url='127.0.0.1:8000/download_gcode_file/' + gcode_file_re.name,
+            gcode_url='/download_gcode_file/' + gcode_file_re.name,
             gcode_safepath=gcode_path,
             gcode_printed='False',
             gcode_selected='False',
@@ -138,14 +162,22 @@ def upload_gcode_file(request):
         # return HttpResponse("uploaded over, print started!")
         return redirect('/list/')
     else:
-        return render(request, 'upload_gcode.html')
+        return redirect('/list/')
 
 
 def print_gcode(request):
     if request.method == 'POST':
         gcode_id = request.POST['gcode_id']
         GcodeFile.objects.filter(gcode_id=gcode_id).update(gcode_selected='True')
-    return redirect('/list/')
+        return redirect('/list/')
+    if request.method == 'GET':
+        gcode_file_sel = GcodeFile.objects.all().first()
+        if gcode_file_sel:
+            gcode_id = model_to_dict(gcode_file_sel)['gcode_id']
+            GcodeFile.objects.filter(gcode_id=gcode_id).update(gcode_selected='True')
+            return redirect('/list/')
+        else:
+            return redirect('/list/')
 
 
 def test(request):
@@ -154,19 +186,18 @@ def test(request):
         print('POST from', printer_state['printer_id'])
 
         # 将打印机状态写入数据库
-        Printer.objects.get_or_create(
-            printer_id=printer_state['printer_id'],
-        )
-
-        Printer.objects.filter(printer_id=printer_state['printer_id']).update(
-            is_operational=printer_state['is_operational'],
-            is_printing=printer_state['is_printing'],
-            is_cancelling=printer_state['is_cancelling'],
-            is_closed_or_error=printer_state['is_closed_or_error'],
-            is_paused=printer_state['is_paused'],
-            is_pausing=printer_state['is_pausing'],
-            is_ready=printer_state['is_ready'],
-        )
+        if RegisteredPrinter.objects.filter(printer_id=printer_state['printer_id']):
+            Printer.objects.filter(printer_id=printer_state['printer_id']).update(
+                is_operational=printer_state['is_operational'],
+                is_printing=printer_state['is_printing'],
+                is_cancelling=printer_state['is_cancelling'],
+                is_closed_or_error=printer_state['is_closed_or_error'],
+                is_paused=printer_state['is_paused'],
+                is_pausing=printer_state['is_pausing'],
+                is_ready=printer_state['is_ready'],
+            )
+        else:
+            return HttpResponse('打印机未注册')
 
         # Printer.objects.all().delete()
 
@@ -188,6 +219,7 @@ def test(request):
                     GcodeFile.objects.filter(gcode_id=gcode_id).update(gcode_selected='False')
                     GcodeFile.objects.filter(gcode_id=gcode_id).update(gcode_printer_id=printer_state['printer_id'])
                     # 应该要remove
+                    print('123')
                     return r
                 except Exception:
                     return Http404('error')
@@ -195,30 +227,3 @@ def test(request):
             return HttpResponse('recieved')
     if request.GET:
         pass
-
-# def register_print(request):
-#     if request.method == 'GET':
-#         return render(request, 'register.html')
-#
-#     if request.method == 'POST':
-#         printer_id = request.POST["请输入printer_id"]
-#         owner = request.POST["请输入owner"]
-#         address = request.POST["请输入adress"]
-#
-#         if not printer_id:
-#             return HttpResponse('printer_id不能为空')
-#         if not owner:
-#             return HttpResponse('密码不能为空')
-#         if not address:
-#             return HttpResponse('确认密码不能为空')
-#         if username and password and repeat_password:
-#             if password == repeat_password:
-#                 # filter() 函数用于过滤序列，过滤掉不符合条件的元素，返回由符合条件元素组成的新列表
-#                 user_project = models.User.objects.filter(username=username).first()
-#                 if user_project:
-#                     return HttpResponse('用户名已存在')
-#                 else:
-#                     models.User.objects.create(username=username, password=password).save()
-#                     return redirect('/login/')
-#             else:
-#                 return HttpResponse('两次输入的密码不一致')
